fixed time bug in max messages fixed wrong temperature calculation and pontential bug in maxWeekProgram message added comments

+++ middleware/src/main/java/org/ambientlight/rfmbridge/messages/max/MaxSetTemperatureMessage.java
@@ -46,7 +46,7 @@ public class MaxSetTemperatureMessage extends MaxMessage implements RequestMessa
 
 
 	public float getTemp() {
-		return (float) (payload[10] & 0x3f) / 2;
+		return (payload[10] & 0x3f) / 2.0f;
 	}
